+++ components/PromptEditor.tsx
@@ -1,4 +1,3 @@
-import { Textarea } from "@mantine/core";
 import { useEffect, useState } from "react";
 
 import { useControlledUncontrolled } from "./useControlledUncontrolled";
@@ -42,13 +41,31 @@ export function PromptEditor(props: PromptEditorProps) {
   }, [promptText, setPrompt]);
 
   return (
-    <Textarea
-      label="prompt"
-      value={promptText}
-      onChange={(evt) => setPromptText(evt.currentTarget.value)}
-      autosize
-      maxRows={10}
-      minRows={3}
-    />
+    <div>
+      <div className="flex gap-2">
+        <label
+          htmlFor="prompt"
+          className="mb-2 block text-sm font-medium text-gray-900 dark:text-white"
+        >
+          prompt
+        </label>
+        <p
+          className="cursor-pointer text-sm text-red-800 hover:text-red-600"
+          onClick={() => {
+            setPromptText("");
+          }}
+        >
+          clear
+        </p>
+      </div>
+      <textarea
+        id="prompt"
+        rows={4}
+        value={promptText}
+        onChange={(evt) => setPromptText(evt.currentTarget.value)}
+        placeholder="Enter your prompt: photograph of a bear"
+        className="block w-full rounded-lg border border-gray-300 bg-gray-50 p-2.5  text-gray-900 focus:border-blue-500 focus:ring-blue-500 dark:border-gray-600 dark:bg-gray-700 dark:text-white dark:placeholder-gray-400 dark:focus:border-blue-500 dark:focus:ring-blue-500"
+      />
+    </div>
   );
 }